Breakout board revision 1.2
- The breakout board is moving back to a much more passive device than
in revision 1.0 and 1.1. The decision was made early for most logic to
be carried out on host FMC board. We need to stick by this decision to
prevent feature creep, maintenance issues, and synchronization bugs
- Removed the OLED display
- Removed rotary encoder
- Teensy 4.0 was replaced with a Tiny FPGA Bx board so that LVDS outputs
can be an ~0 delay shift-registered version of the digital input,
output, and headstage power/connection state, and button pad state
- Headstage status is now provided via a simple through-hole LED next to
each rocker switch