Update omnetics-adapter-64 with Filipe's suggestions
- The suggestions addressed are as follows:
- Increase alls pads of J1 by at least 0.5 mm to the exterior. I believe
this board will be completely assembled by hand and this facilitates
assembly and shorts removal.
- Have 2 mm from the tip of the Omnetics pad to the outline of the
board. This gives us room to apply the necessary epoxy.
- Add some 0.3 holes around the Omnetics to help to fix the epoxy.
- Reduce a little bit the roundness of the Omnetics pads, but not
completely. I would say to half of the current roundness.
- Move the silk indications away from the epoxy application and the
extension of the J1 pads. (done in the board i'm sending)
- Can we write Open-Ephys somewhere in the top silk?
- In the panel, Vcut only for the sides of the board parallel to the
connectors. Milling in the rest. Easier to detach due to J1 being
outside the board boundaries.
- In the panel, a distance between board center to board center of 25 mm
is suggested. (I'm sending the current panel for reference)
- In the panel, if the non-90degrees individual board's corner can
prevail, it would be better: it's more elegant and to grab-friendly.